refactor : 1차 피드백 후 수정 오타 chco -> chocoBody 로 수정 컨벤션 수정 빈칸(띄어쓰기)
