+++ CodeStarterCamp_Week1/iceCream.swift
@@ -1,0 +1,26 @@
+//
+//  iceCream.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by 곽우종 on 2021/12/18.
+//
+
+import Foundation
+
+/* step 2 코드
+printIce(lengthOfIce:8)
+printBar(lengthOfBar:4)
+*/
+
+func printIce(lengthOfIce : Int) {
+    for _ in 1...lengthOfIce{
+        print("***********")
+    }
+}
+
+func printBar(lengthOfBar : Int) {
+    for _ in 1...lengthOfBar{
+        print("    | |")
+    }
+}
+
