+++ CodeStarterCamp_Week1/main.swift
@@ -8,23 +8,7 @@
  
 import Foundation
  
-/* step 2 코드
-printIce(lengthOfIce:8)
-printBar(lengthOfBar:4)
 
-
-func printIce(lengthOfIce : Int) {
-    for _ in 1...lengthOfIce{
-        print("***********")
-    }
-}
-
-func printBar(lengthOfBar : Int) {
-    for _ in 1...lengthOfBar{
-        print("    | |")
-    }
-}
-*/
 
 // stetp 3 코드
 makePaParow(body: "***", topping: " ", size: 12, barSize: 6)
@@ -37,5 +21,12 @@ makePaParow(body: "|0|", topping: " ", size: 12, barSize: 6)
 print("4번째 뺴빼로")
 makePaParow(body: "*_*", topping: "0", size: 12, barSize: 6)
 print("5번째 뺴빼로")
+
+var isTrue = true
+print("\(isTrue)")
+
+
+
+
